decimal ratings bug fix (#36)

+++ src/main/java/com/mas/dashboard/dto/TaskRatingDto.java
@@ -1,7 +1,7 @@
 package com.mas.dashboard.dto;
 
 import lombok.Data;
-
+import java.math.BigDecimal;
 @Data
 public class TaskRatingDto {
   private Long id;
@@ -12,7 +12,7 @@ public class TaskRatingDto {
 
   private Long studentId;
 
-  private float rating;
+  private BigDecimal rating;
 
   private Boolean deleted;
 }
